pyproject.toml
[project]
name = "discordbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiosignal==1.4.0",
    "ascii-magic>=2.3.0",
    "attrs==25.3.0",
    "audioop-lts==0.2.2",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.3",
    "colorama==0.4.6",
    "discord-py>=2.6.3",
    "dotenv==0.9.9",
    "frozenlist==1.7.0",
    "idna==3.10",
    "langdetect==1.0.9",
    "multidict==6.6.4",
    "propcache==0.3.2",
    "pyfiglet==1.0.4",
    "python-dotenv==1.1.1",
    "requests==2.32.5",
    "six==1.17.0",
    "urllib3==2.5.0",
    "yarl==1.20.1",
]
